Update SvelteKit example
