+++ examples/sveltekit/package.json
@@ -21,36 +21,36 @@
     "preview": "vite preview",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
+    "lint": "eslint .",
+    "format": "prettier --write .",
     "transpile": "tsc --watch",
     "flatbread:withnpm": "flatbread start -X npm -- -h"
   },
   "devDependencies": {
     "@flatbread/transformer-yaml": "workspace:*",
-    "@sveltejs/adapter-static": "next",
-    "@sveltejs/kit": "next",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
-    "autoprefixer": "10.4.7",
-    "eslint": "7.32.0",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-svelte3": "3.4.1",
+    "@sveltejs/adapter-static": "^3.0.6",
+    "@sveltejs/kit": "^2.10.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.1",
+    "@types/eslint": "^9.6.1",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.17.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-svelte": "^2.46.0",
     "flatbread": "workspace:*",
-    "graphql": "16.5.0",
-    "prettier": "2.7.1",
-    "prettier-plugin-svelte": "2.4.0",
-    "svelte": "3.49.0",
-    "svelte-check": "2.8.0",
-    "svelte-preprocess": "4.10.7",
-    "tailwindcss": "3.1.6",
-    "tslib": "2.4.0",
-    "typescript": "4.7.4",
-    "vite": "3.0.4",
-    "vite-plugin-transform": "1.1.3"
+    "globals": "^15.13.0",
+    "graphql": "^16.10.0",
+    "prettier": "^3.4.2",
+    "prettier-plugin-svelte": "^3.3.2",
+    "svelte": "^5.16.0",
+    "svelte-check": "^4.1.1",
+    "tailwindcss": "^3.4.17",
+    "tslib": "^2.8.1",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.19.0",
+    "vite": "^5.4.11"
   },
   "type": "module",
   "dependencies": {
-    "svimg": "3.0.0"
+    "svimg": "^3.0.0"
   }
 }
