+++ examples/sveltekit/src/routes/+layout.svelte
