+++ examples/sveltekit/svelte.config.js
@@ -1,16 +1,16 @@
-import preprocess from 'svelte-preprocess';
 import adapter from '@sveltejs/adapter-static';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
   extensions: ['.svelte'],
-  preprocess: [preprocess({})],
+  preprocess: vitePreprocess(),
 
   kit: {
-    adapter: adapter(),
-    prerender: {
-      default: true,
-    },
+    adapter: adapter({
+      strict: false,
+      fallback: 'index.html',
+    }),
   },
 };
