+++ examples/sveltekit/tailwind.config.js
@@ -1,11 +1,8 @@
-const config = {
+/** @type {import('tailwindcss').Config} */
+export default {
   content: ['./src/**/*.{html,js,svelte,ts}'],
-
   theme: {
     extend: {},
   },
-
   plugins: [],
 };
-
-module.exports = config;
